server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "A FastAPI Demo Backend showcasing best practices"
authors = [
    {name = "Wenzhuo Zhao", email = "zslyvain@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["fastapi", "backend", "api", "sqlalchemy", "asyncio", "postgresql"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]
dependencies = [
    "fastapi>=0.101.1,<1.0.0",
    "uvicorn[standard]>=0.23.2,<1.0.0",
    "yarl>=1.9.2,<2.0.0",
    "pydantic-settings>=2.0.3,<3.0.0",
    "SQLAlchemy[asyncio]>=2.0.0,<3.0.0",
    "alembic>=1.11.3,<2.0.0",
    "asyncpg[sa]>=0.28.0,<1.0.0",
    "greenlet>=2.0.2,<3.0.0",
    "ujson>=5.8.0,<6.0.0",
    "email-validator (>=2.3.0,<3.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "flake8",
    "mypy",
    "pre-commit",
    "wemake-python-styleguide",
    "autoflake",
    "black",
    "isort",
    "pytest-cov",
    "anyio",
    "pytest-env",
    "httpx",
    "ruff>=0.12.4,<1.0.0",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^7.0.0"

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # Do not perform function call in argument defaults (FastAPI Depends pattern)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["server"]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
